Split heluauth and helsinki profile integrations
