+++ tests/test_helauth/test_pipeline_additions.py
@@ -7,12 +7,12 @@ from typing import Any, NamedTuple
 import pytest
 
 from tilavarauspalvelu.enums import ReservationNotification
-from tilavarauspalvelu.integrations.helauth.clients import HelsinkiProfileClient
-from tilavarauspalvelu.integrations.helauth.parsers import ssn_to_date
 from tilavarauspalvelu.integrations.helauth.pipeline import (
     migrate_from_tunnistamo_to_keycloak,
     update_user_from_profile,
 )
+from tilavarauspalvelu.integrations.helsinki_profile.clients import HelsinkiProfileClient
+from tilavarauspalvelu.integrations.helsinki_profile.parsers import ssn_to_date
 from tilavarauspalvelu.integrations.sentry import SentryLogger
 from utils.date_utils import local_datetime
 from utils.external_service.errors import ExternalServiceError
